backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.4
beautifulsoup4==4.13.4
Pillow==11.3.0
google-generativeai==0.8.5
selenium==4.34.2
webdriver-manager==4.0.2
